lets try this a bit more... golangish
