+++ server_test.go
@@ -1,11 +1,9 @@
-package server
+package main
 
 import (
 	"encoding/json"
 	"github.com/dgrijalva/jwt-go"
 	"github.com/stretchr/testify/assert"
-	"github.com/tengen-io/server/providers"
-	"github.com/tengen-io/server/test"
 	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
@@ -15,7 +13,7 @@ import (
 )
 
 func TestServer_LoginHandler(t *testing.T) {
-	server := makeServer()
+	server := makeTestServer()
 	reqBody := "{\"email\": \"test1@tengen.io\", \"password\": \"hunter2\"}"
 
 	req, err := http.NewRequest("POST", "/login", strings.NewReader(reqBody))
@@ -39,17 +37,12 @@ func TestServer_LoginHandler(t *testing.T) {
 	assert.NoError(t, err)
 }
 
-func makeServer() *Server {
-	db := test.MakeDb()
+func makeTestServer() *Server {
+	db := MakeTestDb()
 	config := ServerConfig{
 		"", 0, false,
 	}
 	duration, _ := time.ParseDuration("1 week")
-	identityProvider := providers.NewIdentityProvider(db, 1)
-	return NewServer(&config, nil, providers.NewAuthProvider(db, []byte("supersecret"), duration), identityProvider)
-}
-
-
-func TestMain(m *testing.M) {
-	test.Main(m)
+	identityProvider := NewIdentityProvider(db, 1)
+	return NewServer(&config, nil, NewAuthProvider(db, []byte("supersecret"), duration), identityProvider)
 }
